Changed size_t to off_t to handle files >4GB properly; most just needed to include stdio.h
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@39500 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/diskimage.h
@@ -158,6 +158,6 @@ extern void disk_image_attach_log(const disk_image_t *image, signed int lognum,
                                   unsigned int unit, unsigned int drive);
 extern void disk_image_detach_log(const disk_image_t *image, signed int lognum,
                                   unsigned int unit, unsigned int drive);
-extern uint32_t disk_image_size(const disk_image_t *image);
+extern off_t disk_image_size(const disk_image_t *image);
 
 #endif
